+++ noora/plugins/mysql/drop/drop_plugin.py
@@ -27,6 +27,10 @@ class DropPlugin(MysqlPlugin):
         Fail.fail_on_invalid_environment(environment, properties)
         prepared_args['environment'] = environment
 
+        connection_string = arguments.get('connection_string')
+        prepared_args['connection_string'] = property_helper.connection_credentials(connection_string)
+        Fail.fail_on_host_mismatch(host, prepared_args['connection_string'])
+
         alias = arguments.get('alias')
         Fail.fail_on_invalid_alias(alias, properties)
         # if an alias is given, only this database will be installed, other databases will be
@@ -66,16 +70,22 @@ class DropPlugin(MysqlPlugin):
 
         objects = properties.get('drop_objects')
 
-        # retrieve the user credentials for this database project.
-        users = properties.get('mysql_users')
+        # retrieve the database connection credentials through
+        # the commandline option --connection-string
+        connection_string = prepared_args['connection_string']
+        if connection_string:
+            users = connection_string
+        if not connection_string:
+            # retrieve the user credentials for this database project.
+            users = properties.get('mysql_users')
 
-        # try to retrieve the users from the credentials file, when no users are configured in
-        # myproject.json.
-        if not users:
-            # retrieve the name of this database project, introduced in version 1.0.12
-            profile = property_helper.get_profile(properties)
-            if profile:
-                users = profile.get('mysql_users')
+            # try to retrieve the users from the credentials file, when no users are configured in
+            # myproject.json.
+            if not users:
+                # retrieve the name of this database project, introduced in version 1.0.12
+                profile = property_helper.get_profile(properties)
+                if profile:
+                    users = profile.get('mysql_users')
 
         # fail when no users are found. This means that they are not set in myproject.json or
         # credentials.json