+++ noora/plugins/mysql/recreate/__init__.py
@@ -9,8 +9,9 @@ from noora.plugins.mysql.recreate.recreate_plugin import RecreatePlugin
 @click.option('-e', '--environment', required=False, help='Environment')
 @click.option('-a', '--alias', required=False, help='Alias. Overrules the database option')
 @click.option('-v', '--version', required=False, help='The version to recreate')
+@click.option('--connection-string', required=False)
 @click.pass_obj
-def cli(props, host, database, environment, alias, version):
+def cli(props, host, database, environment, alias, version, connection_string):
     """
     Drop and recreate a MySQL database to the specified or latest version.
     """
@@ -19,6 +20,6 @@ def cli(props, host, database, environment, alias, version):
         props,
         {
             'host': host, 'database': database, 'environment': environment,
-            'alias': alias, 'version': version
+            'alias': alias, 'version': version, 'connection_string': connection_string
         },
     )