+++ noora/plugins/mysql/update/__init__.py
@@ -9,8 +9,9 @@ from noora.plugins.mysql.update.update_plugin import UpdatePlugin
 @click.option('-d', '--database', required=False, help='Database name')
 @click.option('-e', '--environment', required=False, help='Environment')
 @click.option('-a', '--alias', required=False, help='Alias. Overrules the database option')
+@click.option('--connection-string', required=False)
 @click.pass_obj
-def cli(props, version, host, database, environment, alias):
+def cli(props, version, host, database, environment, alias, connection_string):
     """
     Update a MSSQL database to the specified version.
     """
@@ -19,6 +20,6 @@ def cli(props, version, host, database, environment, alias):
         props,
         {
             'version': version, 'host': host, 'database': database,
-            'environment': environment, 'alias': alias
+            'environment': environment, 'alias': alias, 'connection_string': connection_string
         },
     )