+++ snippets/libs/create/acme/ddl/cst/comments.sql
@@ -1,1 +1,0 @@
-ALTER TABLE comments ADD CONSTRAINT cmt_eny_fk FOREIGN KEY (eny_id) REFERENCES entries (id);