+++ snippets/libs/create/acme/ddl/idx/referrallinks.sql
@@ -1,2 +1,0 @@
-CREATE INDEX rlk_ste_id_idx ON referrallinks (ste_id);
-CREATE INDEX rlk_usr_id_idx ON referrallinks (usr_id);