+++ test/app/create/orcl/ddl/idx/log.sql
@@ -1,3 +1,0 @@
-CREATE INDEX log_logtype_code_idx ON log (logtype_code);
-CREATE INDEX log_job_name_idx ON log (job_name);
-CREATE INDEX log_unique_session_id_idx ON log (unique_session_id);