+++ test/app/tmp.sql
@@ -1,1 +1,0 @@
-select value into @environment from application_properties where name='application.environment' and value='dev';