+++ test/plugins/mysql/create/orcl/ddl/idx/sites.sql
@@ -1,1 +1,0 @@
-CREATE INDEX ste_delay_unit_idx ON sites (delay_unit);