+++ test/plugins/mysql/create/orcl/ddl/vw/application_properties_v.sql
@@ -1,2 +1,0 @@
-CREATE VIEW application_properties_v AS 
-SELECT id, name, value, description FROM application_properties;