+++ test/plugins/mysql/tmp.sql
@@ -1,1 +1,0 @@
-update application_properties set value='1.0.2' where name='application.version';