+++ test/plugins/test_tripolis_generate_email_plugin.py
@@ -1,7 +1,0 @@
-from org.noora.plugin.tripolis.generate.TripolisGenerateEmailPlugin import TripolisGenerateEmailPlugin
-
-
-plugin = TripolisGenerateEmailPlugin()
-
-moduleName = 'org.noora.plugin.mysql.create.CreatePlugin'
-mod = __import__(moduleName ,globals(), locals(), [''])