+++ tests/plugin/mssql/create/sys/create_schema.sql
@@ -1,4 +1,0 @@
-print "creating schema $(username)"
-USE $(dbf);
-GO
-create schema $(username) authorization $(username)