+++ tests/plugin/mssql/create/sys/create_users.sql
@@ -1,6 +1,0 @@
-print "creating users..."
-:setvar dbf "elsevierdb"
-:setvar username "apps"
-:setvar password "apps"
-:r create_user.sql
-:r create_schema.sql