+++ tests/plugin/mssql/create/sys/drop_schema.sql
@@ -1,6 +1,0 @@
-print "drop schema $(username)"
-drop schema if exists $(username)
-
-print "drop user $(username)"
-drop user $(username)
-drop login $(username)