+++ tests/plugin/mysql/create/acme/dat/application_properties.sql
@@ -1,1 +1,0 @@
-insert into application_properties(name,value) values ('application.domain.name','www.wriggler.nl');