+++ tests/plugin/mysql/create/acme/ddl/cst/entries.sql
@@ -1,1 +1,0 @@
-ALTER TABLE entries ADD CONSTRAINT eny_spe_fk FOREIGN KEY (spe_id) REFERENCES spaces (id);